move the components to type-selector folder
